third_party/blink/web_tests/wpt_internal/app-history/navigate/navigate-info.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  let navInfo = { nav : "info" };
  appHistory.onnavigate = t.step_func_done(e => assert_equals(e.info, navInfo));
  appHistory.navigate("#1", { navigateInfo: navInfo });
}, "navigate() with info");
</script>
